requirements.txt
# Core Flask and API
Flask==3.1.1
flask-cors==6.0.1
Werkzeug==3.1.3

# Authentication
flask-login==0.6.3
flask-sqlalchemy==3.1.1

# Database
SQLAlchemy==2.0.41
peewee==3.18.1
psycopg2-binary==2.9.10 

# Data processing
pandas==2.2.3
numpy==1.26.4
python-dateutil==2.9.0.post0
pytz==2025.2

# Stock data fetching
yfinance==0.2.61
requests==2.32.3
lxml==5.4.0

# Technical indicators
pandas_ta==0.3.14b0

# Utilities
python-dotenv==1.1.0
